refactor: deprecate press props in touchable
